+++ lib/form_obj/form/attributes.rb
@@ -1,0 +1,15 @@
+require 'tree_struct'
+
+module FormObj
+  class Form < ::TreeStruct
+    class Attributes < ::TreeStruct::Attributes
+      def find(name)
+        @items.find { |item| item.name == name.to_sym }
+      end
+
+      def each(&block)
+        @items.each(&block)
+      end
+    end
+  end
+end
